+++ dataclass_wizard/dumpers.py
@@ -1,9 +1,11 @@
 # noinspection PyProtectedMember
+from collections import defaultdict
 from dataclasses import _is_dataclass_instance
 from datetime import datetime, time, date
 from decimal import Decimal
 from enum import Enum
 from typing import Type, Dict, Any, List, Union, Tuple, NamedTupleMeta
+from uuid import UUID
 
 from .abstractions import AbstractDumper
 from .bases import BaseDumpHook
@@ -11,7 +13,7 @@ from .class_helper import (
     dataclass_fields, get_class, dataclass_field_to_json_field)
 from .constants import _DUMP_HOOKS
 from .log import LOG
-from .type_def import NoneType
+from .type_def import NoneType, DD
 from .utils.string_conv import to_camel_case
 
 
@@ -59,6 +61,10 @@ class DumpMixin(AbstractDumper, BaseDumpHook):
         return o.value
 
     @staticmethod
+    def dump_with_uuid(o: UUID, *_):
+        return o.hex
+
+    @staticmethod
     def dump_with_list_or_tuple(
             o: Union[List, Tuple], typ: Type[Union[List, Tuple]],
             dict_factory, hooks):
@@ -79,6 +85,14 @@ class DumpMixin(AbstractDumper, BaseDumpHook):
         return typ((_asdict_inner(k, dict_factory, hooks),
                     _asdict_inner(v, dict_factory, hooks))
                    for k, v in o.items())
+
+    @staticmethod
+    def dump_with_defaultdict(
+            o: Dict, _typ: Type[DD], dict_factory, hooks):
+
+        return {_asdict_inner(k, dict_factory, hooks):
+                _asdict_inner(v, dict_factory, hooks)
+                for k, v in o.items()}
 
     @staticmethod
     def dump_with_decimal(o: Decimal, *_):
@@ -114,9 +128,11 @@ def setup_default_dumper(cls=DumpMixin):
     cls.register_dump_hook(NoneType, cls.dump_with_null)
     # Complex types
     cls.register_dump_hook(Enum, cls.dump_with_enum)
+    cls.register_dump_hook(UUID, cls.dump_with_uuid)
     cls.register_dump_hook(list, cls.dump_with_list_or_tuple)
     cls.register_dump_hook(tuple, cls.dump_with_list_or_tuple)
     cls.register_dump_hook(NamedTupleMeta, cls.dump_with_named_tuple)
+    cls.register_dump_hook(defaultdict, cls.dump_with_defaultdict)
     cls.register_dump_hook(dict, cls.dump_with_dict)
     cls.register_dump_hook(Decimal, cls.dump_with_decimal)
     # Dates and times