+++ dataclass_wizard/type_def.py
@@ -7,12 +7,17 @@ __all__ = [
     'ExplicitNull',
     'NUMBERS',
     'T',
+    'E',
+    'U',
     'M',
+    'DD',
     'N',
     'S'
 ]
 
-from typing import TypeVar, Mapping, Sequence
+from enum import Enum
+from typing import TypeVar, Mapping, Sequence, DefaultDict
+from uuid import UUID
 
 from .constants import PY36, PY38_OR_ABOVE
 from .decorators import discard_kwargs
@@ -25,8 +30,17 @@ NUMBERS = int, float
 # Generic type
 T = TypeVar('T')
 
+# Enum subclass type
+E = TypeVar('E', bound=Enum)
+
+# UUID subclass type
+U = TypeVar('U', bound=UUID)
+
 # Mapping type
 M = TypeVar('M', bound=Mapping)
+
+# DefaultDict type
+DD = TypeVar('DD', bound=DefaultDict)
 
 # Numeric type
 N = TypeVar('N', int, float, complex)