+++ dataclass_wizard/utils/type_conv.py
@@ -2,19 +2,22 @@ __all__ = ['as_bool',
            'as_int',
            'as_str',
            'as_list',
+           'as_enum',
            'as_datetime',
            'as_date',
            'as_time',
            'date_to_timestamp']
 
 from datetime import datetime, time, date
+from enum import Enum
 from numbers import Number
-from typing import Union, List
+from typing import Union, List, Type, AnyStr, Optional
 
-from ..type_def import N, NUMBERS
+from ..errors import ParseError
+from ..type_def import E, N, NUMBERS
 
 
-# What values are considered "truthy" when converted to a boolean type.
+# What values are considered "truthy" when converting to a boolean type.
 # noinspection SpellCheckingInspection
 _TRUTHY_VALUES = ('TRUE', 'T', 'YES', 'Y', '1')
 
@@ -102,6 +105,68 @@ def as_list(o: Union[str, List[str]], sep=','):
         return o
 
     return o.split(sep)
+
+
+def as_enum(o: Union[AnyStr, N],
+            base_type: Type[E],
+            lookup_func=lambda base_type, o: base_type[o],
+            transform_func=lambda o: o.upper().replace(' ', '_'),
+            raise_=True
+            ) -> Optional[E]:
+    """
+    Return `o` if it's already an :class:`Enum` of type `base_type`. If `o` is
+    None or an empty string, return None.
+
+    Otherwise, attempt to convert the object `o` to a :type:`base_type` using
+    the below logic:
+
+        * If `o` is a string, we'll put it through our `transform_func` before
+          a lookup. The default one upper-cases the string and replaces spaces
+          with underscores, since that's typically how we define `Enum` names.
+
+        * Then, convert to a :type:`base_type` using the `lookup_func`. The
+          one looks up by the Enum ``name`` field.
+
+    :raises ParseError: If the lookup for the Enum member fails, and the
+      `raise_` flag is enabled.
+
+    """
+    if isinstance(o, base_type):
+        return o
+
+    if o is None:
+        return o
+
+    if o == '':
+        return None
+
+    key = transform_func(o) if isinstance(o, str) else o
+
+    try:
+        return lookup_func(base_type, key)
+
+    except KeyError:
+
+        if raise_:
+            from inspect import getsource
+
+            enum_cls_name = getattr(base_type, '__qualname__', base_type)
+            valid_values = getattr(base_type, '_member_names_', None)
+            # TODO this is to get the source code for the lambda function.
+            #   Might need to refactor into a helper func when time allows.
+            lookup_func_src = getsource(lookup_func).strip('\n, ').split(
+                'lookup_func=', 1)[-1]
+
+            e = ValueError(
+                f'as_enum: Unable to convert value to type {enum_cls_name!r}')
+
+            raise ParseError(e, o, base_type,
+                             valid_values=valid_values,
+                             lookup_key=key,
+                             lookup_func=lookup_func_src)
+
+        else:
+            return None
 
 
 def as_datetime(o: Union[str, Number, datetime],